Refactoring player paddle code
The prior method was not producing accurate enough hit detection.

I've added in a spherecast inside of the fixed update that checks to see if the ball is within both a certain radius and distance to the paddle. It'll apply a force.impulse to the ball if it detects that it's within its range.

Cleaned out the prior method's variables as I deemed that code obsolete.