New versions of the paddle and dynamic size thrower
This version of the paddle allows VR players to nudge the ball more effectively / accurately than before.
Each client sees the same amount of force be applied to the ball, implemented a oncollisionstay function that appears to be doing its job.

Made a new version of the dynamic sizer, allowing players to throw whichever object they are holding.
There looks to be some sort of discrepancy between non-master clients and master clients on how much force is applied, I will need to look into this later.